+++ metadata/changelog.md
@@ -10,6 +10,7 @@
 
 - Aqueous Soil now pauses auto-harvesting when receiving a redstone signal. Its tooltip has been updated to reflect this.
 - Added a Clay Ball to Clay Block Shaped Recipe.
+- Added a Centrifugal Separator Recipe for Sesame Seeds Crop to Sesame Oil.
 
 ### Fixed